# use b.N instead of fixed value inside benchmark functions