We should use `b.N` as loop index inside benchmark functions for measuring performances exactly.

Before: 

```
$ go test -bench .
BenchmarkGen-4   	1000000000	         0.03 ns/op
PASS
ok  	_/Users/max747/devel/github/gocha	0.170s
```

After:

```
$ go test -bench .
BenchmarkGen-4   	    5000	    269177 ns/op
PASS
ok  	_/Users/max747/devel/github/gocha	1.382s
```